(maint) update installation section to reference multiple git refs
This updates the README section about installing the package from Github to make it clear that several types of git refs can be used, such as a branch name, commit hash, or tag name. This also removes the hardcoded tag name in the example command so that we don't need to keep the example updated as we increment versions.