(maint) update a couple field names and menu item type handling
This makes a few small updates to our data retrieval functions. First, we need to work with recipe externalName instead of name, so this switches over all recipe.name references to recipe.externalName. Second, we need to work with recipe.reconciledNutritionals instead of calculatedNutritionals, so those references are also switched. Finally, we need to know the meal code for menu items (i.e. B1, LV2, etc.). The meal code is represented by the name of the categoryValue where the associated category name is menu item type. So, this adds the name field to the Category and  CategoryValue types so that we pull back that info. Unit tests and mocks are updated to account for all changes included here.