(ENG-823) add more specific types, drop is_perishable support, bump galley-sdk version
This makes a few small adjustments to the recipe data formatting and tests based on issues raised during code review. More specific typing is applied to the function inputs, and the version of galley-sdk is bumped to 0.10.1. In addition, support for the is_perishable category is dropped as the culinary team is no longer planning to use that tag in Galley.