(maint) get standaloneNutrition from subrecipe instead of recipeItem
We are changing how we handle standalone components within Galley and as a result need to access nutrition data for standalone components differently. This updates format_recipe_tree_components_data() to get the value of standaloneNutrition from the standalone subrecipe data layer, as opposed to the standalone recipeItem data layer. Tests and fixtures are updated to account for the change.