(maint) add hasStandalone and baseMeal fields to formatted recipe data
This adds two new fields to the formatted data returned for each recipe. hasStandalone is a boolean field representing whether the recipe includes a standalone component. This new field will allow us to validate on fields specific to the standalone recipe when expected. baseMeal is a categoryValue (tag) being set on each recipe in Galley that we forgot to include in the formatted data previously. Tests are updated to cover the expected behaviors.