(ENG-932) add handling for dynamic standalone serving sizes
This updates get_formatted_recipes_data() and its underlying functions to return a few additional fields about standalone components: standaloneSuggestedServing, standaloneServingSizeWeight, and standaloneServings. This new handling allows us to provide more than one serving of a standalone component within a recipe, and display accurate information to customers on labels and in our website as to how many servings of the standalone component are provided. Previously our implementation assumed that we would only ever provide one serving of standalone components. Tests are added/updated to account for the expected behaviors.