(ENG-1010) remove proof-of-concept code
This removes the POC query and mutation we added while we were setting up galley-sdk and learning how to use the sgqlc tooling. The query, get_recipe_data(), and mutation, add_recipe_instruction(), are not being used and we have no plans to use them in the future. Related tests and Typing entries are also removed.