(ENG-873) retrieve highlight tags from Galley (#46)
* (ENG-873) retrieve highlight tags from Galley

This adds a new field, "highlight_tags", to the response provided by get_formatted_recipes_data(). The highlight_tags field is an array of string values representing what has been assigned to the recipe for the highlight_1 and highlight_2 categories. If no highlight tags have been applied to the recipe, the highlight_tags field is expected to be an empty array.

* (ENG-873) add unit tests for get_recipe_category_tags()

This adds some unit tests to cover the expected behavior of get_recipe_category_tags(), including highlight_tag specific behavior. Mock data is refactored slightly to allow the same data to be used in these unit tests and other related tests.