(ENG-1122) retrieve noNutritionOnWebsite tag (#47)
This retrieves and handles a new tag that is applied to recipes in Galley when we cannot display their nutritional information or ingredient lists on thistle.co. When the tag has been applied to a recipe, the value of the displayNutritionOnWebsite attribute returned by get_formatted_recipe_details() is False. When the tag is not applied to a recipe, the value of displayNutritionOnWebsite defaults to True. Tests are updated to account for the new behavior.