(VEG-34) retrieve more recipe details at menu level (#51)
This updates get_formatted_menu_data() so that it returns the underlying recipe's allergens and highlightTags for each menuItem in the formatted menu. This also fixes a bug that was causing the recipeMenuPhotoUrl attribute on menuItems to always be an empty value, and removes the unused standaloneRecipeId attribute on menuItems. Tests and mock data are updated to confirm the expected behaviors.